pm_qos: Get rid of the allocation in pm_qos_add_request()
All current users of pm_qos_add_request() have the ability to supply
the memory required by the pm_qos routines, so make them do this and
eliminate the kmalloc() with pm_qos_add_request().  This has the
double benefit of making the call never fail and allowing it to be
called from atomic context.

Signed-off-by: James Bottomley <James.Bottomley@suse.de>
Signed-off-by: mark gross <markgross@thegnar.org>
Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>

Conflicts:
	drivers/net/e1000e/netdev.c
	drivers/net/igbvf/netdev.c
	drivers/net/wireless/ipw2x00/ipw2100.c
	include/linux/pm_qos_params.h
	kernel/pm_qos_params.c