rear cam: Enabled tons of debug messages and worked on flash logic.
-It's tough to say what "should be" in this code. We are working with
so much Ad-Hoc code because of the lack of official source for the
camera HAL.

-The s5k5ccgx_set_flash_mode function has been reworked to do 2 things.
 1. Turn the flash ON if the flash_mode is 4.
 2. Turn the flash OFF if the flash_mode is FLASH_OFF.

-flash_mode regardless of value before s5k5ccgx_set_flash_mode is called so
that s5k5ccgx_set_flash can evaluate wether or not to actually turn
on the flash based on what has been set by the camera. ie (on | off | auto)

-Flash during video capture on the rear cam is not correct atm. It should
wait until the record button has been hit before turning on and should turn
off when recording has stopped.

-Debug messages slow down cam a little bit. (will be removed later)